MOONRAY-5367 Record MoonRay affinity status into shared memory to share this info with other MoonRay process (#995)
* MOONRAY-5367 wip

* MOONRAY-5367 wip 2

* MOONRAY-5367 wip 3

* MOONRAY-5367 wip 4

* MOONRAY-5367 wip 5

* MOONRAY-5367 wip 6

* MOONRAY-5367 wip 7

* MOONRAY-5367 wip 8

* MOONRAY-5367 wip 9

* MOONRAY-5367 wip 10

* MOONRAY-5367 wip 11

* MOONRAY-5367 wip 12

* MOONRAY-5367 wip 13

* MOONRAY-5367 wip 14

* MOONRAY-5367 wip 15

* MOONRAY-5637 wip 16

* MOONRAY-5637 wip 17

* MOONRAY-5637 wip 18

* MOONRAY-5637 wip 19

* MOONRAY-5637 wip 20

* MOONRAY-5637 wip 21

* MOONRAY-5367 wip 22

* MOONRAY-5367 wip 23

* MOONRAY-5367 wip 24

* MOONRAY-5367 wip 25

* MOONRAY-5367 wip 26

* MOONRAY-5367 wip 27

* MOONRAY-5367 wip 28

* MOONRAY-5367 wip 29

* MOONRAY-5367 wip 30

* MOONRAY-5367 wip 31

* MOONRAY-5367 wip 32

* MOONRAY-5367 wip 33

* MOONRAY-5367 wip 34

* MOONRAY-=5367 wip 35

* MOONRAY-5367 wip 36

* MOONRAY-5367 wip 37

* MOONRAY-5367 wip 38

* MOONRAY-5367 wip 39

* MOONRAY-5367 wip 40

* MOONRAY-5367 wip 41

* MOONRAY-5367 wip 42

* MOONRAY-5367 wip 43

* MOONRAY-5367 wip 44

* MOONRAY-5367 wip 45

* MOONRAY-5367 wip 46

* MOONRAY-5367 wip 47

* MOONRAY-5367 wip 48

* MOONRAY-5367 wip 49

* MOONRAY-5367 wip 50

* MOONRAY-5367 wip 51

* MOONRAY-5367 wip 52

* MOONRAY-5367 wip 53

* MOONRAY-5367 PR1 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR1 revert to MOONRAY-5367-main1

* MOONRAY-5367 PR1 relocate some files and clean up

* MOONRAY-5367 PR1 cleanup

* MOONRAY-5367 PR1 addressed feedback

* MOONRAY-5367 PR1 addressed feedback 1

* MOONRAY-5367 PR2 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR2 revert to MOONRAY-5367-main2

* MOONRAY-5367 PR2 Shared memory access base class update + minor clean up

* MOONRAY-5367 PR2 additional clean up

* MOONRAY-5367 PR2 addressed review feedback

* MOONRAY-5367 PR2 revert to original MOONRAY-5367-main2 equivarent 2

* MOONRAY-5367 PR3 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR3 revert to MOONRAY-5367-main3

* MOONRAY-5367 PR3 minor clean up only

* MOONRAY-5367 PR3 addressed review feedback

* MOONRAY-5367 PR3 revert to original MOONRAY-5367-main3 equivarent 2

* MOONRAY-5367 PR4 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR4 revert to MOONRAY-5367-main4

* MOONRAY-5367 PR4 affinity info shared memory management

* MOONRAY-5367 PR4 addressed review feedback

* MOONRAY-5367 PR4 minor cleanup 2

* MOONRAY-5367 pre PR5 wip

* MOONRAY-5367 pre PR5 wip 2

* MOONRAY-5367 pre PR5 wip 3

* MOONRAY-5367 pre PR5 wip 4

* MOONRAY-5367 PR5 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR5 revert to MOONRAY-5367-main5

* MOONRAY-5367 PR5 changed to unique_ptr + UnitTest

* MOONRAY-5367 PR6 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR6 revert to MOONRAY-5367-main6

* MOONRAY-5367 PR6 Affinity core resource allocation logic

* MOONRAY-5367 PR6 addressed feedback

* MOONRAY-5367 PR7 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR7 revert to MOONRAY-5367-main7

* MOONRAY-5367 PR7 UnitTest for affinity resource control + Process check utility

* MOONRAY-5367 PR7 addressed feedback

* MOONRAY-5367 PR8 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR8 revert to MOONRAY-5367-main8

* MOONRAY-5367 PR8 added AffinityMapTable

* MOONRAY-5367 PR8 addressed review feedback

* MOONRAY-5367 PR9 revert to MOONRAY-5367-PR

* MOONRAY-5367 PR9 AffinityMapTable UnitTest + AffinityMapTool

* MOONRAY-5367 PR9 cleanup

* MOONRAY_5367 PR9 final clean up

---------

+++ lib/render/util/ThreadPoolExecutor.cc
@@ -1,6 +1,5 @@
-// Copyright 2024 DreamWorks Animation LLC
+// Copyright 2024-2025 DreamWorks Animation LLC
 // SPDX-License-Identifier: Apache-2.0
-
 #include "CpuAffinityMask.h"
 #include "ThreadPoolExecutor.h"
 
@@ -331,7 +330,7 @@ ThreadPoolExecutor::testBootShutdown()
                     nanosleep(&tm, NULL); // yield CPU resources
                 }
 
-                sum += threadId;
+                sum += static_cast<int>(threadId);
             });
     }
